ISPN-5298 HotRod millisecond precision for lifespan/maxidle

+++ client/hotrod-client/src/main/java/org/infinispan/client/hotrod/impl/operations/ReplaceOperation.java
@@ -1,8 +1,10 @@
 package org.infinispan.client.hotrod.impl.operations;
 
 import net.jcip.annotations.Immutable;
+
 import org.infinispan.client.hotrod.Flag;
 import org.infinispan.client.hotrod.impl.protocol.Codec;
+import org.infinispan.client.hotrod.impl.protocol.InternalFlag;
 import org.infinispan.client.hotrod.impl.transport.Transport;
 import org.infinispan.client.hotrod.impl.transport.TransportFactory;
 
@@ -20,7 +22,7 @@ public class ReplaceOperation extends AbstractKeyValueOperation<byte[]> {
 
    public ReplaceOperation(Codec codec, TransportFactory transportFactory,
             byte[] key, byte[] cacheName, AtomicInteger topologyId,
-            Flag[] flags, byte[] value, int lifespan, int maxIdle) {
+            Flag[] flags, byte[] value, long lifespan, long maxIdle) {
       super(codec, transportFactory, key, cacheName, topologyId, flags, value, lifespan, maxIdle);
    }
